Fixes #2 Change variable type

+++ src/Main.java
@@ -1,13 +1,13 @@
 public class Main {
     public static void main(String[] args) {
 
-        int widthOfRoom = 5;
-        int lengthOfRoom = 6;
+        double widthOfRoom = 5.0;
+        double lengthOfRoom = 6.0;
         double heightOfRoom = 2.75;
         double widthOfWallpaper = 1.06;
         int lengthOfRoll = 10;
 
-        int perimeter = 2 * (widthOfRoom + lengthOfRoom);
+        double perimeter = 2 * (widthOfRoom + lengthOfRoom);
 
         double neededNumberOfPanels = (int) Math.round(perimeter / widthOfWallpaper);
